Change verb tense of Vector method names
E.g. changed `add` to `plus`, and `sub` to `minus`, because
these are method which perform vector operations on `this`
and a parameter and return the output, so this tense is
more natural. At some point we may include static methods
performing these operations on their parameters; those static
methods will be `add`, `subtract`, etc.