Simplify AutoMovementCommand to only use forceStoppedController and to use Robot.stopAutoMovement by default
The dual-functionality before, where by default it would use a local
boolean to manage force-stop (and could *optionally* be passed a BoolBox
like Robot.stopAutoMovement which would persistently record force-stop),
led to confusion:

RotateDegreesGyroCommand and DriveForwardEncodersCommand used the
local boolean (because, presumably for lack of thought, their
constructors called super() and did not pass Robot.stopAutoMovement).
When one of those commands was force-stopped, it could run again
because there was no persistent tracking of force-stopping.

When we tested commands that did pass super(Robot.stopAutoMovement, ...),
like RotateToLiftCommand, we did not include a provision for resetting
Robot.stopAutoMovement, and spent time during testing figuring this
out.

Basically, the option to use a local boolean for force-stopping was
one we don't need and which creates confusion.

This commit also adds a clearer println to GyroRotationalCommand and
a similar one to EncoderDrivingCommand, indicating when the command
is force-stopped *in initialize()* (the lack printlns there was
another source of confusion during testing).