Refactor auto-movement commands for modularity
Rather than exposing instance variables to subclasses (by
making them protected), and having setInitialInchesToMove
and setDesiredAngle mutate those variables, the auto-movement
commands now use getInitialInchesToMove and getDesiredAngle,
which *return* the amount to move.

If either get* method returns a value less than 0.1 (in abs value), the
command is cancelled (previously there was a provision to cancel
GyroRotationalCommand if the value was below 0.001).